Move home to router folder in order to match its functionality.
