+++ front-end/src/routes/home/home.component.jsx
@@ -1,10 +1,10 @@
-import './Home.css'
-import Slider from './components/slideshow/Slider'
-import Row from './components/highest-rated-classes-row/Row'
-import BottomNavigationBar from "./routes/bottom-navigation-bar/bottom-navigation-bar.component";
-import SubjectButtons from './components/subject-buttons/SubjectButtons'
+import './home.styles.css'
+import Slider from '../../components/slideshow/Slider'
+import Row from '../../components/highest-rated-classes-row/Row'
+import BottomNavigationBar from "../bottom-navigation-bar/bottom-navigation-bar.component";
+import SubjectButtons from '../../components/subject-buttons/SubjectButtons'
 import Badge from 'react-bootstrap/Badge';
-import FindClassByMajor from "./components/find-class-by-major/find-class-by-major.component";
+import FindClassByMajor from "../../components/find-class-by-major/find-class-by-major.component";
 
 function Home() {
   return (
@@ -20,7 +20,6 @@ function Home() {
       <Row />
       <FindClassByMajor />
       <BottomNavigationBar />
-      
     </>
   );
 }
